requirements.txt
pyswisseph==2.10.03
ephem==4.1.4
numpy==1.24.3
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0 